Python-MySQL/requirement.txt
mysql-connector-python==9.1.0
setuptools==75.2.0
wheel==0.44.0
